item: multiple inheritance for the item class
Since the item class is used now to manage two things:
1. the item record
2. the item circulation transactions

Two base classes are created for each section above in
api.record:ItemRecord, api.circulation:ItemCirculation
The main Item class inherites from these two classes.
