+++ rero_ils/modules/items/api/__init__.py
@@ -1,0 +1,28 @@
+# -*- coding: utf-8 -*-
+#
+# RERO ILS
+# Copyright (C) 2019 RERO
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as published by
+# the Free Software Foundation, version 3 of the License.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+# GNU Affero General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with this program. If not, see <http://www.gnu.org/licenses/>.
+
+"""Item data module."""
+
+from .api import Item, ItemsIndexer, ItemsSearch, item_id_fetcher, \
+    item_id_minter
+from .circulation import ItemCirculation
+from .record import ItemRecord
+
+__all__ = (
+    'Item', 'ItemRecord', 'ItemCirculation', 'ItemsSearch', 'ItemsIndexer',
+    'item_id_fetcher', 'item_id_minter'
+)
