+++ tests/api/test_availability.py
@@ -21,7 +21,8 @@ from invenio_accounts.testutils import login_user_via_session
 from utils import get_json, postdata
 
 from rero_ils.modules.holdings.api import Holding
-from rero_ils.modules.items.api import Item, ItemStatus
+from rero_ils.modules.items.api import Item
+from rero_ils.modules.items.models import ItemStatus
 from rero_ils.modules.items.views import item_availability_text
 from rero_ils.modules.loans.api import LoanAction
 
@@ -261,7 +262,7 @@ def test_item_holding_document_availability(
         class locale:
             language = 'en'
     with mock.patch(
-        'rero_ils.modules.items.api.current_i18n',
+        'rero_ils.modules.items.api.circulation.current_i18n',
         current_i18n
     ):
         end_date = item.get_item_end_date()
