+++ tests/api/test_items_in_transit.py
@@ -21,7 +21,8 @@
 from invenio_accounts.testutils import login_user_via_session
 from utils import postdata
 
-from rero_ils.modules.items.api import Item, ItemStatus
+from rero_ils.modules.items.api import Item
+from rero_ils.modules.items.models import ItemStatus
 from rero_ils.modules.loans.api import Loan, LoanAction
